Update plugins.rst
I was looking for info on how to extend classes from plugins but couldn't find any in the book. I figured it out but it wasn't obvious to me, so I decided to improve this page of the docs a little so that others in the same situation won't have to figure it out via trial and error like I did.